refactoring: moving files related to drawer to the appropriate folder
