+++ localization/react-intl/src/app/components/drawer/DrawerArticles.json
@@ -1,0 +1,27 @@
+[
+  {
+    "id": "articlesComponent.articles",
+    "description": "The navigation name of the articles section",
+    "defaultMessage": "Articles"
+  },
+  {
+    "id": "articlesComponent.claimAndFactChecks",
+    "description": "Label for a list displayed on the left sidebar that includes items that have claim & fact-checks",
+    "defaultMessage": "Claim & Fact-Checks"
+  },
+  {
+    "id": "articlesComponent.explainers",
+    "description": "Label for a list displayed on the left sidebar that includes items that have explainers",
+    "defaultMessage": "Explainers"
+  },
+  {
+    "id": "projectsComponent.importedReports",
+    "description": "Label for a list displayed on the left sidebar that includes items from the 'Imported fact-checks' channel",
+    "defaultMessage": "Imported"
+  },
+  {
+    "id": "projectsComponent.published",
+    "description": "Label for a list displayed on the left sidebar that includes items that have published reports",
+    "defaultMessage": "Published"
+  }
+]
