+++ localization/react-intl/src/app/components/drawer/DrawerFeeds.json
@@ -1,0 +1,22 @@
+[
+  {
+    "id": "projectsComponent.pendingInvitationFeedTitle",
+    "description": "Tooltip for a navigation item that has a status of a pending invitation",
+    "defaultMessage": "Pending invitation: {feedTitle}"
+  },
+  {
+    "id": "projectsComponent.sharedFeedNavHeader",
+    "description": "The navigation name of the shared feeds section with included Beta messaging and the total count of items in the list below",
+    "defaultMessage": "Shared Feeds [{feedsLength}] <sup>BETA</sup>"
+  },
+  {
+    "id": "projectsComponent.newSharedFeed",
+    "description": "Label for the button that navigates to shared feed creation page",
+    "defaultMessage": "New Shared Feed"
+  },
+  {
+    "id": "projectsComponent.noSharedFeeds",
+    "description": "Displayed under the shared feed header when there are no feeds in it",
+    "defaultMessage": "No shared feeds"
+  }
+]
