+++ src/app/components/drawer/DrawerArticles.js
@@ -4,12 +4,12 @@ import Relay from 'react-relay/classic';
 import { Link } from 'react-router';
 import { FormattedMessage } from 'react-intl';
 import cx from 'classnames/bind';
-import ArticleCoreListCounter from '../../article/ArticleCoreListCounter';
-import NewArticleButton from '../../article/NewArticleButton';
-import PublishedIcon from '../../../icons/fact_check.svg';
-import FileDownloadIcon from '../../../icons/file_download.svg';
-import BookIcon from '../../../icons/book.svg';
-import styles from './Projects.module.css';
+import ArticleCoreListCounter from '../article/ArticleCoreListCounter';
+import NewArticleButton from '../article/NewArticleButton';
+import PublishedIcon from '../../icons/fact_check.svg';
+import FileDownloadIcon from '../../icons/file_download.svg';
+import BookIcon from '../../icons/book.svg';
+import styles from './Projects/Projects.module.css';
 
 const DrawerArticlesComponent = ({ team }) => {
   // Get/set which list item should be highlighted
