+++ src/app/components/drawer/DrawerNavigationComponent.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import DrawerNavigationComponent from './DrawerNavigationComponent';
-import DrawerProjects from './Projects';
-
 import { mountWithIntl, getStore } from '../../../../test/unit/helpers/intl-test';
+import DrawerProjects from '.';
+
 
 describe('<DrawerNavigationComponent />', () => {
   const currentUser = {
