+++ src/app/components/drawer/DrawerTipline.js
@@ -6,32 +6,31 @@ import { FormattedMessage, injectIntl } from 'react-intl';
 import { withRouter, Link } from 'react-router';
 import Collapse from '@material-ui/core/Collapse';
 import cx from 'classnames/bind';
-import ProjectsListItem from './ProjectsListItem';
-import NewProject from './NewProject';
-import ProjectsCoreListCounter from './ProjectsCoreListCounter';
-import ButtonMain from '../../cds/buttons-checkboxes-chips/ButtonMain';
-import CreateMedia from '../../media/CreateMedia';
-import Tooltip from '../../cds/alerts-and-prompts/Tooltip';
-import AddIcon from '../../../icons/add_filled.svg';
-import CategoryIcon from '../../../icons/category.svg';
-import ExpandLessIcon from '../../../icons/chevron_down.svg';
-import ExpandMoreIcon from '../../../icons/chevron_right.svg';
-import SharedFeedIcon from '../../../icons/dynamic_feed.svg';
-import InboxIcon from '../../../icons/inbox.svg';
-import LightbulbIcon from '../../../icons/lightbulb.svg';
-import ListIcon from '../../../icons/list.svg';
-import PersonIcon from '../../../icons/person.svg';
-import UnmatchedIcon from '../../../icons/unmatched.svg';
-import Can from '../../Can';
-import DeleteIcon from '../../../icons/delete.svg';
-import ReportIcon from '../../../icons/report.svg';
-import { withSetFlashMessage } from '../../FlashMessage';
-import { assignedToMeDefaultQuery } from '../../team/AssignedToMe';
-import { suggestedMatchesDefaultQuery } from '../../team/SuggestedMatches';
-import { unmatchedMediaDefaultQuery } from '../../team/UnmatchedMedia';
-import { tiplineInboxDefaultQuery } from '../../team/TiplineInbox';
-import styles from './Projects.module.css';
-
+import ProjectsListItem from './Projects/ProjectsListItem';
+import NewProject from './Projects/NewProject';
+import ProjectsCoreListCounter from './Projects/ProjectsCoreListCounter';
+import ButtonMain from '../cds/buttons-checkboxes-chips/ButtonMain';
+import CreateMedia from '../media/CreateMedia';
+import Tooltip from '../cds/alerts-and-prompts/Tooltip';
+import AddIcon from '../../icons/add_filled.svg';
+import CategoryIcon from '../../icons/category.svg';
+import ExpandLessIcon from '../../icons/chevron_down.svg';
+import ExpandMoreIcon from '../../icons/chevron_right.svg';
+import SharedFeedIcon from '../../icons/dynamic_feed.svg';
+import InboxIcon from '../../icons/inbox.svg';
+import LightbulbIcon from '../../icons/lightbulb.svg';
+import ListIcon from '../../icons/list.svg';
+import PersonIcon from '../../icons/person.svg';
+import UnmatchedIcon from '../../icons/unmatched.svg';
+import Can from '../Can';
+import DeleteIcon from '../../icons/delete.svg';
+import ReportIcon from '../../icons/report.svg';
+import { withSetFlashMessage } from '../FlashMessage';
+import { assignedToMeDefaultQuery } from '../team/AssignedToMe';
+import { suggestedMatchesDefaultQuery } from '../team/SuggestedMatches';
+import { unmatchedMediaDefaultQuery } from '../team/UnmatchedMedia';
+import { tiplineInboxDefaultQuery } from '../team/TiplineInbox';
+import styles from './Projects/Projects.module.css';
 
 const DrawerTiplineComponent = ({
   currentUser,
